pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "critik"
version = "0.4.0"
description = "AI-powered code security scanner — catch what Copilot ships and Snyk overcharges for"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Alex LaGuardia", email = "alex@alexlaguardia.dev" }]
keywords = ["security", "scanner", "vibe-coding", "ai", "copilot", "cursor"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

[project.scripts]
critik = "critik.cli:main"

[project.optional-dependencies]
ai = ["groq>=1.0"]
rules = ["pyyaml>=6.0"]
dev = ["pytest>=7.0", "groq>=1.0", "pyyaml>=6.0"]

[project.urls]
Homepage = "https://critik.dev"
Repository = "https://github.com/AlexlaGuardia/critik"
Issues = "https://github.com/AlexlaGuardia/critik/issues"

[tool.hatch.build]
include = ["critik/**"]

[tool.hatch.build.targets.wheel]
packages = ["critik"]

[tool.pytest.ini_options]
testpaths = ["tests"]
